+++ KlamathIrrigationDistrict/DataLayer/DataModels/Customers.cs
@@ -47,10 +47,10 @@ namespace KlamathIrrigationDistrict.DataLayer.DataModels
         public string Lateral { get; set; }
 
         //this number holds the entirety of allotment
-        public decimal TotalAllotment { get; set; }
+        public float TotalAllotment { get; set; }
 
         //this number will change given the user's request
-        public decimal CurrentAllotment { get; set; }
+        public float CurrentAllotment { get; set; }
 
         public int Ride { get; set; }
